requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-socketio==5.1.1
opencv-python==4.5.3.56
numpy==1.21.2
onnxruntime==1.8.1
python-dotenv==0.19.0
psutil==5.8.0
boto3==1.18.44
botocore==1.21.44
python-engineio>=4.0.0
python-socketio>=5.0.0